requirements.txt
python-telegram-bot[job-queue]>=20.0
google-cloud-speech>=2.21.0
google-generativeai>=0.3.0
pydub>=0.25.1
python-dateutil>=2.8.2
dateparser>=1.1.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
audioread>=3.0.0
elevenlabs>=1.0.0
aiohttp>=3.9.0
